backend/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
icalendar==5.0.11
recurring-ical-events==2.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
